VNC-Server/Win32: also grab layered windows
Per default, the screen grab is done via BitBlt() and the SRCCOPY flag
which has the disadvantage of not grabbing layered (usually translucent)
windows. Before each grab loop through all windows and look whether
there're layered ones. If so, additionally use the CAPTUREBLT flag for
BitBlt().

This intentionally breaks the invisible-windows-feature of Anti-iTALC.