DemoServerSlave: change user role upon DemoServer start
When acting as DemoServerSlave the global user role in ItalcCore::role
is set to RoleOther. However authentication of the demo server against
the local VNC server usually will not then because RoleOther has no keys
set up. Therefore require the DemoServerMaster to also send the role of
the user using IMA.