ItalcCore/ItalcVncConnection: QWaitCondition-based sleeping
When sleeping between framebuffer updates / connection retries, do not
use QThread::msleep() as we do not always have a chance then to quit
the thread properly in time. Instead use a QWaitCondition object and
wait on a mutex with the update interval as timeout. When stopping
we can simply call QWaitCondition::wakeAll() which will immediately
return from the wait() call.