ICA/ItalcCoreServer: allow connection from localhost with role != other
Usually we do not allow allow connections to the VNC server if it is
running with a role other than ItalcCore::RoleOther (i.e. prevent
teachers from being watched without permission). However for the demo
server to work we need to accept the connection in this case - do this
by checking whether the connection comes from localhost.