ItalcCore/ItalcCoreConnection: check ICC pointer retrieved via ICC-Tag
Under rare circumstances, no ItalcCoreConnectionTag might be set in an
rfbClient structure (i.e. when attaching an ItalcCoreConnection to an
ItalcVncConnection after the latter one has been established a
connection). Check this case explicitely so we do not dereference a
NULL pointer. Returning false instead ensures, that the connection is
being closed and opened again.