ItalcCore/ICA: renamed DisplayLock to ScreenLock and added InputLock slave
To make the screen lock functionality being named consistently through
the whole software, references to "display lock" have been renamed to
reflect the screen lock functionality. Furthermore added a basic input
lock functionality which does almost the same as the screen lock except
it grabs a screenshot of the desktop and shows it instead of the lock
symbol.