ItalcCore/Win32: kill wisptis.exe at startup
For an unknown reason, there's a problem when building iTALC with recent
MinGW/GCC/Qt toolchains. As soon as ICA is running as service and the
user starts one of iTALC's applications (IMA, Setup) the mouse cursor
freezes. This only happens if wisptis - a Microsoft Tablet PC support
software - is installed an running. As a bad hack, we now kill all
instances of wisptis.exe at startup until there's a better solution or
the bug in the build toolchain has been fixed.