ItalcCore/Configuration/XmlStore: properly load and save UTF-8 strings
We have to use QDomDocument::toByteArray() instead of
QDomDocument::toString() when writing an XML file from a QDomDocument.
For formal reasons also specify the encoding in the first line even
though it's not neccessarily needed for loading the document properly
with QDomDocument.

Fixes loading and saving configuration values with real UTF-8 content.