IMC: rewrote code for reading and writing the LogonACL registry value
Management of the LogonACL registry value was completely broken and
needed to be rewritten. QSettings is no longer used for reading and
writing the LogonACL value as it does not properly support binary
values. Furthermore the internal logic applying/resetting etc. has
been rewritten. Last but not least the LogonACL is now stored base64
encoded in XML files and also applied properly when using the
"-ApplySettings" command line option.