Ipc::QtSlaveLauncher: process events while waiting for slave to finish
When stopping a slave, we want the whole application still to be
responsive and for example send queued messages. We therefore must not
use QProcess::waitForFinished(). Instead implement an own wait loop
which calls QCoreApplication::processEvents().

Fixes the bug of improperly terminated slaves (e.g. SystemTrayIconSlave)
on ICA service shutdown and thus heavily reduces the time needed to stop
the service.