IMA: revised all dialogs
Revised all dialogs in IMA to be more consistent and lightweight
regarding both user interface and implementation. The client settings
dialog has been simplified and the VPN tab has been removed as the
according functionality still is not available. The CmdInputDialog
has been reimplemented as RunCommandsDialog using the Qt Designer.