ICA/DemoServer: fixed qtcpsocketDispatcher and framebuffer update logic
The qtcpsocketDispatcher read and wrote wrong data when the first call
to read()/write() did not read/write all requested data. This has been
fixed by respecting the number of bytes read/written. Furthermore do
not use QTcpSocket::waitForBytesWritten() but QTcpSocket::flush().

The framebuffer update logic has improved in such a way that updates
will not be sent periodically but only on request. This is achieved
by connecting the readyRead() signal of the socket to
DemoServer::processClient().