IMC: add ICA firewall exception for all firewall profiles
On Windows >= Vista the firewall has been advanced such that it has
various distinct profiles (private, work, public) with different rules.
With the previous code base we added a firewall rule for ICA for the
currently active profile which might not be desired. Instead enable the
rule for all available profiles.

Closes #3156332.