ICA/WindowsService: give service a start to chance before stopping it
On Windows XP we sometimes restart the service twice if on the one hand
a WTS_SESSION_LOGOFF event occurs and on the other hand the session ID
changes (when using fast user switch). As the time interval between
these restarts is too small for the server to properly start up and
listen for shutdown events, it missed these events in the past. In such
situations send the shutdown event multiple times and wait a bit after
stopping the service.

Closes #3163302.