ICA, IMA, IMC: use MinGW ANSI stdio support on Win32
According to http://sourceforge.net/apps/trac/mingw-w64/wiki/_vswprintf%20missing
calls to vswprintf are implemented as calls to _vswprintf() which
is not implemented in various older versions of msvcrt.dll and thus iTALC
does not work on these Windows installations.

Thanks to Sönke Schwardt-Krummrich for pointing out this issue.