libvncserver/sockets: always redefine EWOULDBLOCK and ETIMEDOUT on Win32
Commit b71f01bf693aa162226751aca26c972089d051a5 introduced a severe
regression which completely broke socket read and write functions for
libvncserver on Win32. The symbols EWOULDBLOCK and ETIMEDOUT always have
to correspond to the WSA error codes on Win32 in order to make the code
in rfbReadExactTimeout() and rfbWriteExact() work properly

This finally fixes the broken demo mode. Thanks to the Univention GmbH
for sponsoring the debugging of this issue.