Core: AccessControlProvider: always identify groups by distinguished name
For better distinction between local and LDAP groups, always use the
distinguished name of an LDAP group as identifying name which is used
when defining and checking access control rules. Maybe we can make this
configurable later as there might be situations where a generic
configuration is deployed to different schools and therefore the teacher
groupd should be identified by its environment-independent common name.