Core: ItalcVncConnection: stop thread termination timer on thread finish
If the ItalcVncConnection thread is finishing properly we have to stop
the termination timer as it might fire after the thread has been started
again. This happens when hiding and showing a client window in master
application within a short period and makes the client inaccessible
until restarting the application.