IMA: ClassroomManager: do not update client widgets manually
The client widgets do know best on their own when to update/repaint.
Therefore remove the update mechanism in ClassroomManager which dates
back to times when image files were fetched from client computers
manually instead of using the RFB protocol.

This greatly reduces the number of UI updates and thus the overall
CPU load of the master application.