Implement new naming scheme for all software components
To avoid confusion especially with the terms "client", "service",
"server" and "master" we introduce a new and consistent naming scheme
for all software components as well as source directories:

  * core: iTALC Core
  * service: iTALC Service
  * master: iTALC Master
  * configurator: iTALC Configurator

All binaries are prefixed with "italc-" followed by the component name:

  * italc-core[.exe]
  * italc-service[.exe]
  * italc-master[.exe]
  * italc-configurator[.exe]

The Windows service is now named "ItalcService" instead of "icas".

Installing this new version of iTALC requires a complete removal of
previous iTALC installations in order to properly clean up files which
do not exist any longer.