Service: WindowsService: listen to WTS_SESSION events again
Only monitoring session IDs is not sufficient anymore as session IDs
do not necessarily change if a user logs on. This hasn't been a problem
so far however now we evaluate the local state (i.e. hostname and logged
on user) on server start and use this information e.g. for listening
on localhost only. If the server is not restarted when user logs on/off
(because the session ID doesn't change) we have an inconsistent
behaviour which could also lead to security problems.