Plugins: Demo: DemoServer: rewrite protocol and thread handling
Instead of using blocking reads/writes use a state machine and event
handling for establishing the connection to the demo client. Also do not
inherit DemoServerClient from QThread but use a standard QThread object
and change the affinity of the DemoServerClient object. This makes the
demo mode finally work properly on Windows where e.g. blocking socket
reads do not work as expected.