Master: ComputerManager: find current room by matching host names only
Instead of resolving each host in the network object directory to an
IP address and match it against local IP addresses use the information
as is and compare directly against local host name or local IP
addresses. This saves DNS lookups for each host which lead to massive
startup times in larger environments with several hundreds of hosts.