Core: VeyonVncConnection: do not terminate thread on Windows
Thread termination on Windows causes deadlock when calling any QThread
member functions such as isRunning() or the destructor if the thread
itself is stuck in a blocking (e.g. network) function. Therefore do not
terminate the thread on Windows but let it run in background as long
as the blocking function (which will eventually timeout somewhen) is
running.