Plugins: Demo: rewrite demo server as an efficient VNC proxy
The previous demo server implementation was a complete VNC server which
sent changed regions from a framebuffer updated by an own VNC client
connection. This was laggy, inefficient and unreliable.

The demo server now acts as a special kind of VNC proxy by connecting
directly to the local VNC server and save all framebuffer update
messages into an internal queue. All demo clients are served from this
queue so we don't need to identify and compress changed data for each
client but use the far better heuristics in the VNC server.