Master: manage ComputerControlInterfaces in model
The ComputerManager is only responsible for fetching data from the
NetworkObjectDirectory and provide models for room and computer
selection. The corresponding instances of ComputerControlInterfaces
no longer are managed by the ComputerManager. Instead the renamed
ComputerControlListModel uses the computer objects providede by the
ComputerManager to create and manage ComputerControlInterface instances.
This allows to clean up the complex signal/slot connections for model
updates.