Core: VeyonVncConnection: revise main loop
This fixes timing issues inside the main loop by decreasing sleep time
by the time already elapsed in the loop cycle. Also full updates are not
requested any longer and instead incremental updates are requested in
every loop cycle. This improves response and update times for thumbnails
in the Master.