Core: VncConnection: simplify connection shutdown
Whenever we want to shutdown the connection it won't be used afterwards
any longer at all. Therefore provide an explicit method for stopping and
deleting later as well as a method which only tells the thread to stop.

This also removes the thread termination timer as it doesn't work
properly on all platforms anyway. Resources will be freed as soon as
a blocking network function inside the connection thread times out and
the thread terminates properly itself.