LdapDirectory: add computer search scope handling
When using containers/OUs as locations computer objects are not located
directly below the configured computer DN and are never found if
recursive search is disabled. In such cases access control (which
in the end calls computersByHostName() when retrieving locations of
a computer) would fail even though recursive search is not required
and thus disabled.

Closes #327.