Demo: implement screen selection and management
This new feature allows the user to select which screen(s) to share by
telling the demo client to only display the parts of the framebuffer
belonging to the selected screen(s).

The implementation is still very basic and lacks support for selecting
e.g. 2 out of 3 screens. Also the selected screen should be rememebered
between program starts.