ScreenshotManagementPanel: refactor directory updates
QFilesystemModel only supports local directories, i.e. it fails to
update properly when the screenshot directory is located in a roaming
profile or on a network share.

Instead the list model is now populated manually. Updates are triggered
by both QFilesystemWatcher (which seems to support non-local dirs)
and the new Filesystem::screenshotDirectoryModified() signal.

Closes #668.