LinuxUserFunctions: use listSessions() in isAnyUserLoggedOn()
Querying systemd sessions is much more reliable than calling "who" and
also allows distinguishing between graphical user sessions and all other
types of sessions.

This fixes the processing of the access control rule condition
"No user logged on" on Linux which evaluated to false when a user opens
a local console (tty) session. While this behaviour by itself is
correct, it's not what's intended by access control rules relying on
this condition. Often such a rule is configured to always allow access
to the local graphical display manager. If a console session is opened,
access to that graphical display manager should still be possible.