DesktopAccessDialog: use uniform button layout
The layout (order) of buttons in dialogs varies across platforms,
toolkits and desktop environments. For a consistent user experience on
Linux where a Qt-based display manager such as SDDM and Gtk-based
sessions such as Gnome/Xfce may be used, use a fixed button box layout
with Yes/Always on the left and No/Never on the right.