ComputerControlServer: avoid blocking in updateTrayIconToolTip()
When performing a reverse lookup for the addresses of the connected
peers, this can block for quite a while, especially in case of missing
reverse PTR records in the DNS server.

This new implementation starts populating the tooltip with IP addresses
and updates the tooltip later as soon as the FQDNs have been resolved.

(cherry picked from commit d522d80909eee2a7307f821945d7b5f1b6d6f8db)