Core: refactor ownerships and destruction of connections
Now that ComputerControlInterface is the only user of VeyonConnection,
we can make VeyonConnection have a VncConnection instance internally
instead of being attached to.

Additionally refactor the destruction mechanisms and order since it did
not work properly with dynamically created and destructed event loops
(e.g. when used in thread pools via QtConcurrent). Connecting the
deleteMethod() with receiver `this` eventually did not call the
deleteMethod() if the thread or event loop for the object disappeared
in the meanwhile.

(cherry picked from commit a9accaec63c816ddcf629630cf40d28d5610e87e)