fix press and hold and maximized state of ComputerZoomWidget on windows os
The ComputerZoomWidget window was not working correctly on windows OS’s. The reason was, that when a new window gets created on windows systems one or more (but in my tests not more than three) mouse move events are for unknown reasons generated, even while the mouse is perfectly still.

As mouseMoveEvents are one of the conditions that are checked and result in ending the feature execution, the window closed itself, before the user could get the chance to look at it.

To solve this problem a counter variable was introduced, which allows up to three mouse move events to happen, before it is deemed a valid user decision to end the feature.

As a second improvement this commit stops the mousePressAndHold timer when starting the feature, as keeping it running could trigger unforeseen problems.

The third improvement fixes the size and positioning of the ComputerZoomWidget window to the screen size of the controller. This is achieved by maximizing the window.  Using this method the rescaling of the vnc connection is triggered and will make target device screens completely visible even if the resolution of controller pc is smaller or the target devices have multiple screens.