Demo: implement bandwidth agnostic quality control
Now that tight encoding with variable image quality is supported, we
can change the quality dynamically depending on the consumed bandwidth
and the configured bandwidth limit. The default limit is 100 MB/s
which equals a regular Gigabit network. For networks with lower
available bandwidth (e.g. wireless networks), the limit can be adjusted
in the Demo server configuration tab in advanced view.

Closes #484 and #749.