WebApiController/HttpServer: add endpoints for debugging
When the environment variable VEYON_WEBAPI_DEBUG is set, detailed debug
information are now available via /api/v1/debug/info. This replaces the
previous debug messages being logged when receiving SIGUSR1 which on the
one hand only was available on Linux and on the other hand could cause
a deadlock when locking the connections mutex in the signal handler.