NetworkObjectDirectory: optimize propagation of changed objects
Aggregate object changes and propagate them after timeout of 100 ms.
This significantly improves performance of initial population of the
directory, especially when used in conjunction with
NetworkObjectTreeModel and NetworkObjectFilterProxyModel which cause
cascades of rowCount()/parent()/fetchMore() calls. With very large
nested directories this can lead to extreme recursion depths leading
to OOM crashes.