ComputerManager: hide all locations on failed location detection
If the "Show current location only" option is enabled, no locations
should be displayed at all when the location of the local computer
cannot not be determined. This prevents potentially unwanted access
to other locations and computers, which is consistent with the generally
restrictive nature of the "Show current location only" option.

(cherry picked from commit 2bc10e91cd66355f2764157850050bb99215f948)